block: add block package
This package holds common functionality we need across components when
working with blocks
